Move rangeiterator to sequenceutils, move sequenceutils from operator to data dir
